requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
cryptography
python-multipart
pydantic
python-dotenv
